+++ lib/components/Menu/MenuBody.tsx
@@ -1,12 +1,15 @@
 import type { ReactElement} from 'react';
+import { useState } from 'react';
 import React, { useEffect, useRef } from 'react';
 import { Backdrop } from '../Backdrop';
 import type { MenuPosition } from './menu.types';
+import type { Instance } from '@popperjs/core';
 import { createPopper } from '@popperjs/core';
 import styles from './MenuBody.module.scss';
 import { Portal } from '../Portal';
-import type { IListItemProps} from '../List';
+import type { IListItemProps } from '../List';
 import { List } from '../List';
+import { useOnDestroy } from '../../hooks';
 
 export interface IMenuBodyProps {
 	children?: ReactElement<IListItemProps> | ReactElement<IListItemProps>[];
@@ -20,11 +23,12 @@ export interface IMenuBodyProps {
 export const MenuBody = (props: IMenuBodyProps): JSX.Element => {
 
 	const { parentRef, children, className, shadow = true, menuPosition = 'left', onClickBackdrop } = props;
+	const [popperInstance, setPopperInstance] = useState<Instance>();
 	const menuBodyRef: React.RefObject<HTMLDivElement> = useRef<HTMLDivElement>(null);
 
 	useEffect(() => {
 		if (menuBodyRef && parentRef.current && menuBodyRef.current) {
-			createPopper(parentRef.current, menuBodyRef.current, {
+			const popperInstance: Instance = createPopper(parentRef.current, menuBodyRef.current, {
 				placement: `${menuPosition}-start`,
 				modifiers: [
 					{
@@ -45,6 +49,9 @@ export const MenuBody = (props: IMenuBodyProps): JSX.Element => {
 					},
 				]
 			});
+			setPopperInstance(popperInstance);
+		} else {
+			popperInstance?.destroy();
 		}
 	}, [menuBodyRef]);
 
@@ -59,6 +66,10 @@ export const MenuBody = (props: IMenuBodyProps): JSX.Element => {
 	const handleClickBackdrop = (): void => {
 		onClickBackdrop && onClickBackdrop();
 	};
+
+	useOnDestroy(() => {
+		popperInstance?.destroy();
+	});
 
 	return (
 		<Portal className='menu-root'>