+++ lib/components/Tooltip/tooltip.enums.ts
@@ -1,0 +1,6 @@
+export enum TooltipPlacement {
+	top = 'top',
+	bottom = 'bottom',
+	right = 'right',
+	left = 'left'
+}