+++ src/main/examples/TooltipPage.tsx
@@ -3,6 +3,7 @@ import type { ITooltipProps} from 'lib';
 import { Button, FormControl, Tooltip } from 'lib';
 import type { IShowcaseBaseProps} from './components';
 import { withOptions } from './components';
+import { TooltipPlacement } from 'lib/components/Tooltip/tooltip.enums';
 
 export const TooltipPageBase = ({ settingValues, setSettingsControls }: IShowcaseBaseProps<ITooltipProps>): JSX.Element => {
 
@@ -17,7 +18,8 @@ export const TooltipPageBase = ({ settingValues, setSettingsControls }: IShowcas
 					{ label: 'bottom', value: 'bottom' }
 				]
 			}),
-			isOpen: new FormControl(false, [], 'checkbox', { label: 'isOpen', hint: 'weather' }),
+			isOpen: new FormControl(false, [], 'checkbox', { label: 'isOpen', hint: '' }),
+			isShowClose: new FormControl(false, [], 'checkbox', { label: 'isShowClose', hint: '' }),
 			delay: new FormControl(0, [], 'number', { label: 'delay', hint: 'delay for displaying and hiding a tooltip' }),
 		});
 	}, []);
@@ -27,6 +29,7 @@ export const TooltipPageBase = ({ settingValues, setSettingsControls }: IShowcas
 			text={settingValues.text}
 			placement={settingValues.placement}
 			isOpen={settingValues.isOpen}
+			isShowClose={settingValues.isShowClose}
 			delay={settingValues.delay}
 		>
 			<Button>
@@ -38,5 +41,5 @@ export const TooltipPageBase = ({ settingValues, setSettingsControls }: IShowcas
 
 export const TooltipPage: () => JSX.Element = withOptions(TooltipPageBase, {
 	text: 'some tooltip text',
-	placement: 'top'
+	placement: TooltipPlacement.top
 }, 'TooltipPageBase');